refactor(gpt-gateway): migrate service to use proper kafka schema for complete event
