+++ libs/schema-registry/src/lib/ai-conversion-complete/index.ts
