+++ libs/schema-registry/src/lib/ai-conversion-complete/key.ts
