+++ libs/schema-registry/src/lib/ai-conversion-complete/value.ts
